ADD: 상세페이지 레이아웃 구현중
