+++ public/data/productDetailImageList.json
@@ -1,0 +1,18 @@
+[
+  {
+    "id": 1,
+    "imgUrl": "../../public/images/Detail/detail.jpg"
+  },
+  {
+    "id": 2,
+    "imgUrl": "../../public/images/Detail/detail.jpg"
+  },
+  {
+    "id": 3,
+    "imgUrl": "./images/Detail/detail.jpg"
+  },
+  {
+    "id": 4,
+    "imgUrl": "./images/Detail/detail.jpg"
+  }
+]
