+++ src/components/Detail/Detail.scss
@@ -1,0 +1,22 @@
+.datail {
+  .detailNav {
+    color: #606060;
+    font-size: 10px;
+  }
+  .detailMainContainer {
+    .detailImageContainer {
+      .detailImageBox {
+        width: 250px;
+        height: 250px;
+        background-color: aqua;
+        .detailImageEach {
+          width: 100%;
+          height: 100%;
+        }
+      }
+    }
+    .detailDescription {
+      font-size: 22px;
+    }
+  }
+}
