--- src/pages/Detail/Detail.scss
@@ -1,15 +1,0 @@
-.datail {
-  .detailNav {
-    color: #606060;
-    font-size: 10px;
-  }
-  .detailMainContainer {
-    .detailImageContainer {
-      .detai  lImageBox {
-        width: 250px;
-        height: 250px;
-      }
-    }
-    .detailDescription
-  }
-}
